requirements.txt
# Core web framework
flask>=3.0.0
werkzeug>=3.0.0

# Machine Learning & Computer Vision
tensorflow>=2.16.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
scikit-image>=0.22.0

# Scientific computing
numpy>=1.24.0,<2.0.0
scipy>=1.11.0
pillow>=10.0.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.12.0
plotly>=5.18.0

# Image processing
imagehash>=4.3.0
